修复bug

+++ src/main/java/com/wyu/service/FactoryService.java
@@ -8,5 +8,5 @@ public interface FactoryService {
     void add(Factory factory);
     int update(Factory factory);
     List<Factory> searchAll();
-    int deal(int id, boolean flag);
+    int deal(long id, boolean flag);
 }
